Implement Equivalent flag, add unit tests.
Implemented a flag which can be used to make certain characters
equivalent to each other, for instance to treat '_' and '-'
identically.  This could also be used for simple pattern matching.

Refactored how case-insensitive matching works and how flags are
implemented, in light of the above.

Added an error return value to Generate() and GenerateReverse(),
rather than call panic().  We now check the fmt.Fprintln() return
value on the first write, which should catch when an unusable output
file handle is provided.

Finally, added a unit test framework, which generates code in a
temporary directory, executes it, and compares the matcher function
behavior with what's expected.