Add README and LICENSE.
Unicode normalization support is still missing, but I'm considering
this done enough for now.  I'm using it in two different projects: a
parser for a DSL, and to convert IANA-registered charset and language
names to constants in an email tool.  It's feature-complete enough for
both uses, so perhaps other folks will find it useful as well.