HasPrefix/HasSuffix matching; code refactoring
Added support for generating functions which match the beginning or
end of the input string and ignore the rest.

While I was at it, I did a little bit of code refactoring.  I
separated the creation of the state machine from code generation, with
separate passes for each.

The refactoring offsets a little bit of the added complexity from the
partial matching support, but I'm still not particularly happy with
the current readability/understandability of this code.  Specifically,
checkAmbiguity() feels like it might contain dragons.