Allow unlimited match length/complexity
Rather than return an error when we overflow the uint64 used for the
state machine, we now collapse the intermediate values and continue
with a new state machine.  This allows for an unlimited number of
matches of virtually unlimited length and complexity.

As part of this work, I refactored ambiguity detection, including unit
tests.  There still lie a few dragons there, but I think the code is a
lot more readable now.

*ErrAmbiguity is returned for ambiguous matches, and ErrBadFlags is
now a type (instead of a value) for consistency.  N.B. this is a
breaking change if one were checking the return value from Generate
and GenerateReverse.

I added a GenerateTest routine, which generates a simplistic unit test
from known matches.  This is run for each of the code generation
tests.  It's of marginal utility at the moment since it ignores flags,
but I may improve it in the future.