New flags: StopUpon, Ignore, IgnoreExcept
Add some Generate() flags which allow for the construction of more
complicated matchers:

StopUpon specifies runes which are to be treated like a string
boundary, i.e. cause matching to immediately terminate.  Ignore
specifies runes which are to be ignored in the input.  IgnoreExcept is
similar, except all runes not listed will be ignored.  See the godoc
for more information and examples.

To make these flags (and Equivalent) easier to use, I added a helper
function, Range(), and some predefined ranges for ASCII runes:
Numbers, Letters, Uppercase, Lowercase, and Alphanumeric.

There were a handful of minor cleanups along the way, notably some
better tests for runeEquivalents and slightly improved test coverage
of error handling for bad flags.

With these enhancements, code generation (generate.go) has grown
somewhat unwieldy; I plan to clean it up (and speed up the tests) in a
future commit.  The generated code is still about as efficient as I
can make it.